Moved over bulk of provenance code to provenance translator.
