+++ src/Makefile.am
@@ -203,6 +203,8 @@ souffle_sources = \
         ast2ram/Location.h                                 \
         ast2ram/ProvenanceClauseTranslator.cpp             \
         ast2ram/ProvenanceClauseTranslator.h               \
+        ast2ram/ProvenanceTranslator.cpp                   \
+        ast2ram/ProvenanceTranslator.h                     \
         ast2ram/ValueIndex.cpp                             \
         ast2ram/ValueIndex.h                               \
         ast2ram/ValueTranslator.cpp                        \
