+++ src/ast2ram/ProvenanceTranslator.h
@@ -1,0 +1,38 @@
+/*
+ * Souffle - A Datalog Compiler
+ * Copyright (c) 2018 The Souffle Developers. All rights reserved
+ * Licensed under the Universal Permissive License v 1.0 as shown at:
+ * - https://opensource.org/licenses/UPL
+ * - <souffle root>/licenses/SOUFFLE-UPL.txt
+ */
+
+/************************************************************************
+ *
+ * @file ProvenanceTranslator.h
+ *
+ ***********************************************************************/
+
+#pragma once
+
+#include "ast2ram/AstToRamTranslator.h"
+
+namespace souffle::ast2ram {
+
+class ProvenanceTranslator : public AstToRamTranslator {
+public:
+    ProvenanceTranslator() = default;
+    ~ProvenanceTranslator() = default;
+
+protected:
+    void translateProgram(const ast::TranslationUnit& translationUnit) override;
+    void addNegation(ast::Clause& clause, const ast::Atom* atom) override;
+private:
+    /** translate RAM code for subroutine to get subproofs */
+    Own<ram::Statement> makeSubproofSubroutine(const ast::Clause& clause);
+
+    /** translate RAM code for subroutine to get subproofs for non-existence of a tuple */
+    Own<ram::Statement> makeNegationSubproofSubroutine(const ast::Clause& clause);
+
+    void addProvenanceClauseSubroutines(const ast::Program* program);
+};
+}  // namespace souffle::ast2ram
